Add base Websocket class
